requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.2.0
duckdb==0.10.0

# AWS Bedrock
boto3==1.34.34
botocore==1.34.34

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Visualization
plotly==5.18.0
altair==5.2.0

# OpenSearch (for future real-time integration)
opensearch-py==2.4.2

# Type hints
typing-extensions==4.9.0
